+++ processor/src/test/resources/expected/com/example/component_dependency/Arez_ActAsComponentFieldDependencyModel.java
@@ -1,0 +1,106 @@
+package com.example.component_dependency;
+
+import arez.Arez;
+import arez.ArezContext;
+import arez.Component;
+import arez.Disposable;
+import arez.SafeProcedure;
+import arez.component.DisposeNotifier;
+import arez.component.Identifiable;
+import arez.component.internal.ComponentKernel;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+
+@Generated("arez.processor.ArezProcessor")
+public final class Arez_ActAsComponentFieldDependencyModel extends ActAsComponentFieldDependencyModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
+  private static volatile int $$arezi$$_nextId;
+
+  private final ComponentKernel $$arezi$$_kernel;
+
+  public Arez_ActAsComponentFieldDependencyModel() {
+    super();
+    final ArezContext $$arezv$$_context = Arez.context();
+    final int $$arezv$$_id = ( Arez.areNamesEnabled() || Arez.areRegistriesEnabled() || Arez.areNativeComponentsEnabled() ) ? ++$$arezi$$_nextId : 0;
+    final String $$arezv$$_name = Arez.areNamesEnabled() ? "ActAsComponentFieldDependencyModel." + $$arezv$$_id : null;
+    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "ActAsComponentFieldDependencyModel", $$arezv$$_id, $$arezv$$_name, () -> $$arezi$$_nativeComponentPreDispose() ) : null;
+    this.$$arezi$$_kernel = new ComponentKernel( Arez.areZonesEnabled() ? $$arezv$$_context : null, Arez.areNamesEnabled() ? $$arezv$$_name : null, $$arezv$$_id, Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNativeComponentsEnabled() ? null : this::$$arezi$$_preDispose, null, null, true, false, false );
+    if ( null != this.time ) {
+      DisposeNotifier.asDisposeNotifier( this.time ).addOnDisposeListener( this, this::dispose );
+    }
+    this.$$arezi$$_kernel.componentConstructed();
+    this.$$arezi$$_kernel.componentComplete();
+  }
+
+  final int $$arezi$$_id() {
+    return this.$$arezi$$_kernel.getId();
+  }
+
+  @Override
+  @Nonnull
+  public final Integer getArezId() {
+    return $$arezi$$_id();
+  }
+
+  private void $$arezi$$_preDispose() {
+    if ( null != this.time ) {
+      DisposeNotifier.asDisposeNotifier( this.time ).removeOnDisposeListener( this );
+    }
+  }
+
+  private void $$arezi$$_nativeComponentPreDispose() {
+    this.$$arezi$$_preDispose();
+    this.$$arezi$$_kernel.notifyOnDisposeListeners();
+  }
+
+  @Override
+  public void addOnDisposeListener(@Nonnull final Object key, @Nonnull final SafeProcedure action) {
+    this.$$arezi$$_kernel.addOnDisposeListener( key, action );
+  }
+
+  @Override
+  public void removeOnDisposeListener(@Nonnull final Object key) {
+    this.$$arezi$$_kernel.removeOnDisposeListener( key );
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return this.$$arezi$$_kernel.isDisposed();
+  }
+
+  @Override
+  public void dispose() {
+    this.$$arezi$$_kernel.dispose();
+  }
+
+  @Override
+  public final int hashCode() {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      return Integer.hashCode( $$arezi$$_id() );
+    } else {
+      return super.hashCode();
+    }
+  }
+
+  @Override
+  public final boolean equals(final Object o) {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      if ( o instanceof Arez_ActAsComponentFieldDependencyModel ) {
+        final Arez_ActAsComponentFieldDependencyModel that = (Arez_ActAsComponentFieldDependencyModel) o;
+        return $$arezi$$_id() == that.$$arezi$$_id();
+      } else {
+        return false;
+      }
+    } else {
+      return super.equals( o );
+    }
+  }
+
+  @Override
+  public final String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + this.$$arezi$$_kernel.getName() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}