+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Avoid issuing a warning when the `@ComponentDependency` annotation is applied to a field with a type annotated by `@ActAsComponent` or to a method with a return type annotated by `@ActAsComponent`.
+
 ### [v0.144](https://github.com/arez/arez/tree/v0.144) (2019-09-16)
 [Full Changelog](https://github.com/arez/arez/compare/v0.143...v0.144)
 [API Differences](https://arez.github.io/api-diff/?key=arez&old=0.143&new=0.144)